ovs-router: Introduce ovs/rule/show command.
Add a new command `ovs-appctl ovs/rule/show` for printing OVS internal
routing rules database:

  ovs-appctl ovs/rule/show
  32763: from 7.7.7.17 lookup 20
  32764: from all lookup 15
  32765: from 7.7.7.7 lookup 10

The rules are sorted from the highest priority being the first to the
lowest being the last one.

The default OVS routing table is implicitly being the lowest priority
and is not shown in the 'ovs/rule/show' output.

Signed-off-by: Dima Chumak <dchumak@nvidia.com>
Change-Id: Id887a6f6476778fe5cd93d930d444260c3a787eb