router-table: Add back-off to periodic router reset.
OVS periodically polls netlink socket for route and rule updates from
the kernel, and it will trigger full router table reset upon any
relevant change. In the event of a high volume notification updates,
combined with a big overall number of routes and rules, the control
thread may start getting starved spending most of the cycles parsing
router updates:

  wakeup due to [POLLIN] on fd 91 (NETLINK_ROUTE<->NETLINK_ROUTE) at lib/netlink-socket.c:1418 (96% CPU usage)
  wakeup due to [POLLIN] on fd 96 (FIFO pipe:[140158448]) at lib/ovs-rcu.c:259 (96% CPU usage)
  wakeup due to [POLLIN] on fd 96 (FIFO pipe:[140158448]) at lib/ovs-rcu.c:259 (99% CPU usage)
  wakeup due to [POLLIN] on fd 14 (<->/var/run/openvswitch/db.sock) at lib/stream-fd.c:157 (99% CPU usage)
  wakeup due to [POLLIN] on fd 96 (FIFO pipe:[140158448]) at vswitchd/bridge.c:431 (99% CPU usage)
  wakeup due to 105-ms timeout at vswitchd/bridge.c:3195 (99% CPU usage)

Such behavior was triggered for example when adding 500 rules with a
table lookup action, where each rule referenced a unique table. And each
table had 100 routes. In total it amounted in 50k routes in the custom
tables.

To optimize CPU usage under such conditions a back-off mechanism is used
before doing router reset, to reduce the frequency of a full dump of
router tables and rules. The back-off mechanism is dynamic in a sense
that it takes into consideration the duration of the last router reset
operation so the back-off delay is also increasing with a higher volume
of total routes and rules.

Signed-off-by: Dima Chumak <dchumak@nvidia.com>
Change-Id: I0554f1d1e3e0350dcaf08726fe7cd71ab0f6d5ed