ofproto-dpif: Protect group's hash_map.
A group has a hash_map that contains the array of its buckets.
Revalidator threads access it to check if some bucket is alive, while
the main thread may modify the group, re-allocating this array. This race
condition may lead to invalid memory access.

Introduce a rw-lock to take a write-lock while modifying a group and a
read-lock for revalidators.

Issue: 4503873
Fixes: 0a8f6beb54ab ("ofproto-dpif: Fix dp_hash mapping after select group modification.")
Change-Id: I124869c61587863cfff4b85e5a15d01d0cdd480b
Signed-off-by: Eli Britstein <elibr@nvidia.com>
Acked-by: Roi Dayan <roid@nvidia.com>